+++ activerecord/lib/active_record/validations/uniqueness.rb
@@ -24,12 +24,14 @@ module ActiveRecord
             raise UnknownPrimaryKey.new(finder_class, "Can not validate uniqueness for persisted record without primary key.")
           end
         end
-        relation = scope_relation(record, table, relation)
+        scope_hash = scope_relation(record)
+        relation = relation.where(scope_hash) unless scope_hash.empty?
         relation = relation.merge(options[:conditions]) if options[:conditions]
 
         if relation.exists?
           error_options = options.except(:case_sensitive, :scope, :conditions)
           error_options[:value] = value
+          error_options[:scopes] = scope_hash unless scope_hash.empty?
 
           record.errors.add(attribute, :taken, error_options)
         end
@@ -89,18 +91,17 @@ module ActiveRecord
         klass.none
       end
 
-      def scope_relation(record, table, relation)
-        Array(options[:scope]).each do |scope_item|
+      def scope_relation(record)
+        Array(options[:scope]).inject({}) do |hash, scope_item|
           if reflection = record.class._reflect_on_association(scope_item)
             scope_value = record.send(reflection.foreign_key)
             scope_item  = reflection.foreign_key
           else
             scope_value = record._read_attribute(scope_item)
           end
-          relation = relation.where(scope_item => scope_value)
+          hash[scope_item] = scope_value
+          hash
         end
-
-        relation
       end
 
       def map_enum_attribute(klass, attribute, value)